# api/server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "MLX API Extension for Kubeflow Pipelines"
  version: "0.1.29-dont-cache-kfservices"
  title: "MLX API"
basePath: "/apis/v1alpha1"
schemes:
- "http"
paths:
...
  /models:
    get:
      tags:
      - "ModelService"
      operationId: "list_models"
      parameters:
      - name: "page_token"
        in: "query"
        required: false
        type: "string"
      - name: "page_size"
        in: "query"
        required: false
        type: "integer"
        format: "int32"
      - name: "sort_by"
        in: "query"
        description: "Can be format of \"field_name\", \"field_name asc\" or \"field_name\
          \ desc\"\nAscending by default."
        required: false
        type: "string"
        x-nullable: true
      - name: "filter"
        in: "query"
        description: "A string-serialized JSON dictionary with key-value pairs that\
          \ correspond to the Model's attribute names and their respective values\
          \ to be filtered for."
        required: false
        type: "string"
        x-nullable: true
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiListModelsResponse"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
    post:
      tags:
      - "ModelService"
      operationId: "create_model"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/apiModel"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiModel"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/{id}:
    get:
      tags:
      - "ModelService"
      operationId: "get_model"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiModel"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
    delete:
      tags:
      - "ModelService"
      operationId: "delete_model"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "204":
          description: "Deletion successful."
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/{id}/download:
    get:
      tags:
      - "ModelService"
      summary: "Returns the model artifacts compressed into a .tgz (.tar.gz) file."
      operationId: "download_model_files"
      produces:
      - "application/gzip"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "include_generated_code"
        in: "query"
        description: "Include generated run scripts in download"
        required: false
        type: "boolean"
        default: false
        allowEmptyValue: true
      responses:
        "200":
          description: "A TGZ file containing the model artifacts."
          schema:
            type: "file"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/{id}/generate_code:
    get:
      tags:
      - "ModelService"
      operationId: "generate_model_code"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiGenerateModelCodeResponse"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/{id}/run:
    post:
      tags:
      - "ModelService"
      operationId: "run_model"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      - name: "pipeline_stage"
        in: "query"
        description: "pipeline stage, either 'train' or 'serve'"
        required: true
        type: "string"
        enum:
        - "train"
        - "serve"
      - name: "execution_platform"
        in: "query"
        description: "execution platform, i.e. 'kubernetes', 'knative'"
        required: true
        type: "string"
      - name: "run_name"
        in: "query"
        description: "name to identify the run on the Kubeflow Pipelines UI, defaults\
          \ to model identifier"
        required: false
        type: "string"
      - in: "body"
        name: "parameters"
        description: "optional run parameters, must include 'github_url' and 'github_token'\
          \ if credentials are required"
        required: false
        schema:
          $ref: "#/definitions/dictionary"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiRunCodeResponse"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/{id}/templates:
    get:
      tags:
      - "ModelService"
      operationId: "get_model_template"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "string"
      responses:
        "200":
          description: "A successful response."
          schema:
            $ref: "#/definitions/apiGetTemplateResponse"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/upload:
    post:
      tags:
      - "ModelService"
      operationId: "upload_model"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "uploadfile"
        in: "formData"
        description: "The model YAML file to upload. Can be a GZip-compressed TAR\
          \ file (.tgz, .tar.gz) or a YAML file (.yaml, .yml). Maximum size is 32MB."
        required: true
        type: "file"
      - name: "name"
        in: "query"
        required: false
        type: "string"
      responses:
        "201":
          description: ""
          schema:
            $ref: "#/definitions/apiModel"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/upload_from_url:
    post:
      tags:
      - "ModelService"
      operationId: "upload_model_from_url"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "url"
        in: "formData"
        description: "URL pointing to the model YAML file."
        required: true
        type: "string"
      - name: "name"
        in: "query"
        description: "Optional, the name of the model to be created overriding the\
          \ name in the YAML file."
        required: false
        type: "string"
      - name: "access_token"
        in: "formData"
        description: "Optional, the Bearer token to access the 'url'."
        required: false
        type: "string"
      responses:
        "201":
          description: "Model was created successfully."
          schema:
            $ref: "#/definitions/apiModel"
        default:
          description: "Something went wrong."
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/{id}/upload:
    post:
      tags:
      - "ModelService"
      operationId: "upload_model_file"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "The model identifier."
        required: true
        type: "string"
      - name: "uploadfile"
        in: "formData"
        description: "The file to upload, overwriting existing. Can be a GZip-compressed\
          \ TAR file (.tgz), a YAML file (.yaml), Python script (.py), or Markdown\
          \ file (.md)"
        required: true
        type: "file"
      responses:
        "204":
          description: ""
          schema:
            $ref: "#/definitions/apiModel"
        default:
          description: ""
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/featured:
    post:
      tags:
      - "ModelService"
      operationId: "set_featured_models"
      parameters:
      - in: "body"
        name: "model_ids"
        description: "Array of model IDs to be featured."
        required: true
        schema:
          type: "array"
          items:
            type: "string"
      responses:
        "200":
          description: "Success"
        default:
          description: "Something went wrong"
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
  /models/publish_approved:
    post:
      tags:
      - "ModelService"
      operationId: "approve_models_for_publishing"
      parameters:
      - in: "body"
        name: "model_ids"
        description: "Array of model IDs to be approved for publishing."
        required: true
        schema:
          type: "array"
          items:
            type: "string"
      responses:
        "200":
          description: "Success"
        default:
          description: "Something went wrong"
          schema:
            $ref: "#/definitions/apiStatus"
      x-swagger-router-controller: "swagger_server.controllers.model_service_controller"
...
definitions:
...
  apiAsset:
    type: "object"
    required:
    - "description"
    - "name"
    properties:
      id:
        type: "string"
      created_at:
        type: "string"
        format: "date-time"
      name:
        type: "string"
      description:
        type: "string"
      featured:
        type: "boolean"
      publish_approved:
        type: "boolean"
      related_assets:
        type: "array"
        items:
          type: "string"
      filter_categories:
        type: "object"
        additionalProperties:
          type: "string"
    description: "Abstract asset with properties shared across all MLX asset types (except pipelines)."
    example:
      id: "05au-84nr-1tey-0qwr"
      created_at: "2019-01-23T04:56:07.000+00:00"
      name: "Image Recognition Model"
      description: "A machine learning model to identify objects in images"
      filter_categories:
        domain: image-recognition
        platform:
          - kubernetes
          - kfserving
        language: python
...
  apiGenerateModelCodeResponse:
    type: "object"
    properties:
      scripts:
        type: "array"
        description: "The source code samples to use the model in a pipeline"
        items:
          $ref: "#/definitions/apiModelScript"
    example:
      scripts:
      - pipeline_stage: "train"
        execution_platform: "wml"
        script_code: "TODO: python script example"
      - pipeline_stage: "serve"
        execution_platform: "kubernetes"
        script_code: "TODO: python script example"
  apiGetTemplateResponse:
    type: "object"
    properties:
      template:
        type: "string"
        description: "The YAML template file content"
      url:
        type: "string"
        description: "The URL to download the template text from S3 storage (Minio)"
    example:
      template: "TODO: YAML file content"
      url: "http://min.io:9000/mlpipeline/components/1lBnSAce8lAXP97Yb11eqSk/train_wml.yaml"
...
  apiListModelsResponse:
    type: "object"
    properties:
      models:
        type: "array"
        items:
          $ref: "#/definitions/apiModel"
      total_size:
        type: "integer"
        format: "int32"
      next_page_token:
        type: "string"
    example:
      models:
      - id: "max-audio-classifier"
        created_at: "2019-04-23T04:56:07.000+00:00"
        name: "Max Audio Classifier"
        description: "Identify sounds in short audio clips"
        labels:
          domain: "speech recognition"
        framework:
          name: "tensorflow"
          version: "1.8"
          runtimes:
            name: "python"
            version: "3.5"
        trainable: false
        servable: true
      - id: "max-audio-classifier"
        created_at: "2019-04-23T04:56:07.000+00:00"
        name: "Max Audio Classifier"
        description: "Identify sounds in short audio clips"
        labels:
          domain: "speech recognition"
        framework:
          name: "tensorflow"
          version: "1.8"
          runtimes:
            name: "python"
            version: "3.5"
        trainable: false
        servable: true
      next_page_token: "next_page_token"
      total_size: 0
...
  apiModel:
    allOf:
    - $ref: "#/definitions/apiAsset"
    - type: "object"
      required:
      - "framework"
      properties:
        domain:
          type: "string"
        labels:
          type: "object"
          additionalProperties:
            type: "string"
        framework:
          $ref: "#/definitions/apiModelFramework"
        trainable:
          type: "boolean"
        trainable_tested_platforms:
          type: "array"
          items:
            type: "string"
        trainable_credentials_required:
          type: "boolean"
        trainable_parameters:
          type: "array"
          items:
            $ref: "#/definitions/apiParameter"
        servable:
          type: "boolean"
        servable_tested_platforms:
          type: "array"
          items:
            type: "string"
        servable_credentials_required:
          type: "boolean"
        servable_parameters:
          type: "array"
          items:
            $ref: "#/definitions/apiParameter"
    description: "A model that can be trained and/or served."
    example:
      id: "max-audio-classifier"
      created_at: "2019-04-23T04:56:07.000+00:00"
      name: "Max Audio Classifier"
      description: "Identify sounds in short audio clips"
      labels:
        domain: "speech recognition"
      framework:
        name: "tensorflow"
        version: "1.8"
        runtimes:
          name: "python"
          version: "3.5"
      trainable: false
      servable: true
  apiModelFramework:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
      version:
        type: "string"
      runtimes:
        $ref: "#/definitions/apiModelFramework_runtimes"
    example:
      name: "tensorflow"
      version: "1.5"
      runtimes:
        name: "python"
        version: "3.5"
  apiModelScript:
    type: "object"
    required:
    - "execution_platform"
    - "pipeline_stage"
    - "script_code"
    properties:
      pipeline_stage:
        type: "string"
        description: "pipeline stage that this code sample applies to, either 'train'\
          \ or 'serve'"
      execution_platform:
        type: "string"
        description: "execution platform that this code sample applies to, i.e. 'kubernetes',\
          \ 'knative'"
      script_code:
        type: "string"
        description: "the source code to run the model in a pipeline stage"
    example:
      pipeline_stage: "serve"
      execution_platform: "kubernetes"
      script_code: "TODO: python script example"
...
  apiParameter:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
      description:
        type: "string"
      default:
        $ref: "#/definitions/anyValue"
      value:
        $ref: "#/definitions/anyValue"
    example:
      name: "name"
      value: "value"
...
  apiRunCodeResponse:
    type: "object"
    required:
    - "run_url"
    properties:
      run_url:
        type: "string"
        description: "The URL path to view the run details in the Kubeflow Pipeline\
          \ Web UI"
      run_output_location:
        type: "string"
        description: "The location of the output artifacts produced by the run, if\
          \ any."
    example:
      run_url: "/runs/details/0e2680ec-5d5a-11e9-9911-5ed556129669"
...
  apiStatus:
    type: "object"
    properties:
      error:
        type: "string"
      code:
        type: "integer"
        format: "int32"
      details:
        type: "array"
        items:
          $ref: "#/definitions/protobufAny"
...
  anyValue:
    description: "A non-null value of any type like string, number, boolean, array,\
      \ object, see https://stackoverflow.com/a/43328994/5601796 (swagger-2-0-what-schema-to-accept-any-complex-json-value)"
  dictionary:
    type: "object"
    description: "A dictionary of key-value pairs with values of any type."
    example:
      some_key: "some value"
      another_key: true
      some_number: 42
    additionalProperties:
      $ref: "#/definitions/anyValue"
  protobufAny:
    type: "object"
    properties:
      type_url:
        type: "string"
        description: "TODO"
      value:
        type: "string"
        format: "byte"
        description: "Must be a valid serialized protocol buffer of the above specified\
          \ type."
        pattern: "^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$"
        x-regex: "^(?:[A-Za-z0-9+\\/]{4})*(?:[A-Za-z0-9+\\/]{2}==|[A-Za-z0-9+\\/]{3}=)?$"
        x-modifiers: []
    description: "TODO"
  apiModelFramework_runtimes:
    type: "object"
    properties:
      name:
        type: "string"
      version:
        type: "string"